get rid of ares_nowarn casting helpers.  Now that proper datatypes are used, they are no longer necessary.

+++ src/lib/ares_fds.c
@@ -28,7 +28,6 @@
 #include "ares_setup.h"
 
 #include "ares.h"
-#include "ares_nowarn.h"
 #include "ares_private.h"
 
 int ares_fds(ares_channel channel, fd_set *read_fds, fd_set *write_fds)
@@ -65,5 +64,5 @@ int ares_fds(ares_channel channel, fd_set *read_fds, fd_set *write_fds)
     }
   }
 
-  return (int)nfds;
+  return nfds;
 }
